[MOVE/#52] search 폴더 이동
