+++ app/src/main/java/com/napzak/market/presentation/explore/search/state/SearchUiState.kt
@@ -1,4 +1,4 @@
-package com.napzak.market.presentation.search.state
+package com.napzak.market.presentation.explore.search.state
 
 import com.napzak.market.core.common.state.UiState
 import com.napzak.market.domain.genre.model.Genre
